Revert "Merge branch 'development' into release"
This reverts commit 4e4f9c4047626cf378e7bd8e5f2cfd26c2dd28b4, reversing
changes made to 824a0e86e33f845c5541c492fb0ade3f8c3205e7.

+++ activity/src/main/java/com/kairos/persistence/model/staff_settings/StaffActivitySetting.java
@@ -27,13 +27,13 @@ public class StaffActivitySetting extends MongoBaseEntity {
     private LocalTime earliestStartTime;
     private LocalTime latestStartTime;
     private LocalTime maximumEndTime;
-    private List<BigInteger> dayTypeIds;
+    private List<Long> dayTypeIds;
     private LocalTime defaultStartTime;
     private TimeTypeEnum secondLevelTimtype;
 
     public StaffActivitySetting(Long staffId, BigInteger activityId, Long employmentId, Long unitId,
                                 Short shortestTime, Short longestTime, Integer minLength, Integer maxThisActivityPerShift,
-                                boolean eligibleForMove, LocalTime earliestStartTime, LocalTime latestStartTime, LocalTime maximumEndTime, List<BigInteger> dayTypeIds, LocalTime defaultStartTime,TimeTypeEnum secondLevelTimtype) {
+                                boolean eligibleForMove, LocalTime earliestStartTime, LocalTime latestStartTime, LocalTime maximumEndTime, List<Long> dayTypeIds, LocalTime defaultStartTime,TimeTypeEnum secondLevelTimtype) {
         this.staffId = staffId;
         this.activityId = activityId;
         this.employmentId = employmentId;
