+++ kcommons/src/main/java/com/kairos/dto/user/employment/PositionDTO.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigInteger;
 
 /**
  * Created by yatharth on 19/4/18.
@@ -14,6 +13,6 @@ import java.math.BigInteger;
 public class PositionDTO {
     @NotNull
     private String endDate;
-    private BigInteger reasonCodeId;
+    private Long reasonCodeId;
     private Long accessGroupIdOnPositionEnd;
 }
